feat: enhance experience bookings controller with user authentication; update CRUD methods to associate bookings with authenticated users

+++ app/controllers/experience_bookings_controller.ts
@@ -1,19 +1,42 @@
+/* eslint-disable @typescript-eslint/naming-convention */
 import type { HttpContext } from '@adonisjs/core/http'
 import ExperienceBooking from '#models/experience_booking'
 
 export default class ExperienceBookingsController {
-  async index({ response }: HttpContext) {
+  async index({ response, auth }: HttpContext) {
     try {
-      const bookings = await ExperienceBooking.all()
-      return bookings
+      const user = auth.user
+
+      if (!user) {
+        return response.unauthorized({ message: 'User not found' })
+      }
+      const bookings = await ExperienceBooking.query().where('user_id', user.id)
+      return response.ok(bookings)
     } catch (error) {
       return response.internalServerError({ message: 'Error fetching bookings', error })
     }
   }
 
-  async store({ request, response }: HttpContext) {
+  async store({ request, response, auth }: HttpContext) {
     try {
-      const data = request.only(['experience_id', 'user_id', 'status', 'price'])
+      const user = auth.user
+      if (!user) {
+        return response.unauthorized({ message: 'User not found' })
+      }
+
+      const { experience_id, status, price, voucher } = request.only([
+        'experience_id',
+        'status',
+        'price',
+        'voucher',
+      ])
+      const data = {
+        user_id: user.id,
+        experience_id: experience_id,
+        status,
+        price,
+        voucher,
+      }
       const booking = await ExperienceBooking.create(data)
 
       return response.created(booking)
@@ -22,19 +45,35 @@ export default class ExperienceBookingsController {
     }
   }
 
-  async show({ params, response }: HttpContext) {
+  async show({ params, response, auth }: HttpContext) {
     try {
-      const booking = await ExperienceBooking.findOrFail(params.id)
+      const user = auth.user
+      if (!user) {
+        return response.unauthorized({ message: 'User not found' })
+      }
+
+      const booking = await ExperienceBooking.query()
+        .where('user_id', user.id)
+        .where('id', params.id)
+        .firstOrFail()
+
       return booking
     } catch (error) {
       return response.notFound({ message: 'Booking not found' })
     }
   }
 
-  async update({ params, request, response }: HttpContext) {
+  async update({ params, request, response, auth }: HttpContext) {
     try {
-      const booking = await ExperienceBooking.findOrFail(params.id)
-      const data = request.only(['status', 'price'])
+      const user = auth.user
+      if (!user) {
+        return response.unauthorized({ message: 'User not found' })
+      }
+      const booking = await ExperienceBooking.query()
+        .where('user_id', user.id)
+        .where('id', params.id)
+        .firstOrFail()
+      const data = request.only(['status', 'price', 'voucher'])
       booking.merge(data)
       await booking.save()
 
@@ -44,9 +83,17 @@ export default class ExperienceBookingsController {
     }
   }
 
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ params, response, auth }: HttpContext) {
     try {
-      const booking = await ExperienceBooking.findOrFail(params.id)
+      const user = auth.user
+      if (!user) {
+        return response.unauthorized({ message: 'User not found' })
+      }
+      const booking = await ExperienceBooking.query()
+        .where('user_id', user.id)
+        .where('id', params.id)
+        .firstOrFail()
+
       await booking.delete()
 
       return response.noContent()
